Discount Double Added bug fixed

+++ app/src/main/java/com/humaclab/selliscope/model/AddNewOrder.java
@@ -17,7 +17,7 @@ public class AddNewOrder implements Serializable {
         @SerializedName("outlet_id")
         public int outletId;
         @SerializedName("discount")
-        public int discount;
+        public Double discount;
         @SerializedName("products")
         public List<Product> products;
 
@@ -63,7 +63,7 @@ public class AddNewOrder implements Serializable {
                 @SerializedName("outlet_id")
                 public int outlet_id;
                 @SerializedName("discount")
-                public int discount;
+                public Double discount;
             }
         }
     }
